+++ aws-java-sdk-mediaconvert/src/main/java/com/amazonaws/services/mediaconvert/model/AudioDescription.java
@@ -48,6 +48,12 @@ public class AudioDescription implements Serializable, Cloneable, StructuredPojo
 
     private AudioCodecSettings codecSettings;
     /**
+     * Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language code.
+     * The language specified will be used when 'Follow Input Language Code' is not selected or when 'Follow Input
+     * Language Code' is selected but there is no ISO 639 language code specified by the input.
+     */
+    private String customLanguageCode;
+    /**
      * Indicates the language of the audio output track. The ISO 639 language specified in the 'Language Code' drop down
      * will be used when 'Follow Input Language Code' is not selected or when 'Follow Input Language Code' is selected
      * but there is no ISO 639 language code specified by the input.
@@ -275,6 +281,52 @@ public class AudioDescription implements Serializable, Cloneable, StructuredPojo
     }
 
     /**
+     * Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language code.
+     * The language specified will be used when 'Follow Input Language Code' is not selected or when 'Follow Input
+     * Language Code' is selected but there is no ISO 639 language code specified by the input.
+     * 
+     * @param customLanguageCode
+     *        Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language
+     *        code. The language specified will be used when 'Follow Input Language Code' is not selected or when
+     *        'Follow Input Language Code' is selected but there is no ISO 639 language code specified by the input.
+     */
+
+    public void setCustomLanguageCode(String customLanguageCode) {
+        this.customLanguageCode = customLanguageCode;
+    }
+
+    /**
+     * Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language code.
+     * The language specified will be used when 'Follow Input Language Code' is not selected or when 'Follow Input
+     * Language Code' is selected but there is no ISO 639 language code specified by the input.
+     * 
+     * @return Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language
+     *         code. The language specified will be used when 'Follow Input Language Code' is not selected or when
+     *         'Follow Input Language Code' is selected but there is no ISO 639 language code specified by the input.
+     */
+
+    public String getCustomLanguageCode() {
+        return this.customLanguageCode;
+    }
+
+    /**
+     * Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language code.
+     * The language specified will be used when 'Follow Input Language Code' is not selected or when 'Follow Input
+     * Language Code' is selected but there is no ISO 639 language code specified by the input.
+     * 
+     * @param customLanguageCode
+     *        Specify the language for this audio output track, using the ISO 639-2 or ISO 639-3 three-letter language
+     *        code. The language specified will be used when 'Follow Input Language Code' is not selected or when
+     *        'Follow Input Language Code' is selected but there is no ISO 639 language code specified by the input.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public AudioDescription withCustomLanguageCode(String customLanguageCode) {
+        setCustomLanguageCode(customLanguageCode);
+        return this;
+    }
+
+    /**
      * Indicates the language of the audio output track. The ISO 639 language specified in the 'Language Code' drop down
      * will be used when 'Follow Input Language Code' is not selected or when 'Follow Input Language Code' is selected
      * but there is no ISO 639 language code specified by the input.
@@ -476,6 +528,8 @@ public class AudioDescription implements Serializable, Cloneable, StructuredPojo
             sb.append("AudioTypeControl: ").append(getAudioTypeControl()).append(",");
         if (getCodecSettings() != null)
             sb.append("CodecSettings: ").append(getCodecSettings()).append(",");
+        if (getCustomLanguageCode() != null)
+            sb.append("CustomLanguageCode: ").append(getCustomLanguageCode()).append(",");
         if (getLanguageCode() != null)
             sb.append("LanguageCode: ").append(getLanguageCode()).append(",");
         if (getLanguageCodeControl() != null)
@@ -518,6 +572,10 @@ public class AudioDescription implements Serializable, Cloneable, StructuredPojo
             return false;
         if (other.getCodecSettings() != null && other.getCodecSettings().equals(this.getCodecSettings()) == false)
             return false;
+        if (other.getCustomLanguageCode() == null ^ this.getCustomLanguageCode() == null)
+            return false;
+        if (other.getCustomLanguageCode() != null && other.getCustomLanguageCode().equals(this.getCustomLanguageCode()) == false)
+            return false;
         if (other.getLanguageCode() == null ^ this.getLanguageCode() == null)
             return false;
         if (other.getLanguageCode() != null && other.getLanguageCode().equals(this.getLanguageCode()) == false)
@@ -547,6 +605,7 @@ public class AudioDescription implements Serializable, Cloneable, StructuredPojo
         hashCode = prime * hashCode + ((getAudioType() == null) ? 0 : getAudioType().hashCode());
         hashCode = prime * hashCode + ((getAudioTypeControl() == null) ? 0 : getAudioTypeControl().hashCode());
         hashCode = prime * hashCode + ((getCodecSettings() == null) ? 0 : getCodecSettings().hashCode());
+        hashCode = prime * hashCode + ((getCustomLanguageCode() == null) ? 0 : getCustomLanguageCode().hashCode());
         hashCode = prime * hashCode + ((getLanguageCode() == null) ? 0 : getLanguageCode().hashCode());
         hashCode = prime * hashCode + ((getLanguageCodeControl() == null) ? 0 : getLanguageCodeControl().hashCode());
         hashCode = prime * hashCode + ((getRemixSettings() == null) ? 0 : getRemixSettings().hashCode());