+++ aws-java-sdk-mediaconvert/src/main/java/com/amazonaws/services/mediaconvert/model/LanguageCode.java
@@ -15,8 +15,8 @@ package com.amazonaws.services.mediaconvert.model;
 import javax.annotation.Generated;
 
 /**
- * Code to specify the language, following the specification
- * "ISO 639-2 three-digit code":http://www.loc.gov/standards/iso639-2/
+ * Specify the language, using the ISO 639-2 three-letter code listed at
+ * https://www.loc.gov/standards/iso639-2/php/code_list.php.
  */
 @Generated("com.amazonaws:aws-java-sdk-code-generator")
 public enum LanguageCode {