+++ aws-java-sdk-mediaconvert/src/main/java/com/amazonaws/services/mediaconvert/model/transform/CaptionDescriptionMarshaller.java
@@ -29,6 +29,8 @@ public class CaptionDescriptionMarshaller {
 
     private static final MarshallingInfo<String> CAPTIONSELECTORNAME_BINDING = MarshallingInfo.builder(MarshallingType.STRING)
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("captionSelectorName").build();
+    private static final MarshallingInfo<String> CUSTOMLANGUAGECODE_BINDING = MarshallingInfo.builder(MarshallingType.STRING)
+            .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("customLanguageCode").build();
     private static final MarshallingInfo<StructuredPojo> DESTINATIONSETTINGS_BINDING = MarshallingInfo.builder(MarshallingType.STRUCTURED)
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("destinationSettings").build();
     private static final MarshallingInfo<String> LANGUAGECODE_BINDING = MarshallingInfo.builder(MarshallingType.STRING)
@@ -53,6 +55,7 @@ public class CaptionDescriptionMarshaller {
 
         try {
             protocolMarshaller.marshall(captionDescription.getCaptionSelectorName(), CAPTIONSELECTORNAME_BINDING);
+            protocolMarshaller.marshall(captionDescription.getCustomLanguageCode(), CUSTOMLANGUAGECODE_BINDING);
             protocolMarshaller.marshall(captionDescription.getDestinationSettings(), DESTINATIONSETTINGS_BINDING);
             protocolMarshaller.marshall(captionDescription.getLanguageCode(), LANGUAGECODE_BINDING);
             protocolMarshaller.marshall(captionDescription.getLanguageDescription(), LANGUAGEDESCRIPTION_BINDING);